Lunasia 1.0/main/requirements.txt
requests>=2.31.0
openai>=1.0.0
PyQt5>=5.15.0
azure-cognitiveservices-speech>=1.34.0
pygame>=2.5.0
aiohttp>=3.8.0
pandas>=1.5.0
numpy>=1.21.0
psutil>=5.9.0
Pillow>=9.0.0
PyMuPDF>=1.20.0
pytesseract>=0.3.10
